Use imports instead of eval in prerender worker
